moves tz to mixin. applies mixin to views
